Turn a Node object into predictor and outcome.

Turn a Node object into predictor and outcome.

node_to_df(x, drop_last_level = TRUE)

Arguments

x

Node object

drop_last_level

TRUE unused

Value

a named list of x and y, being respectively the predictor data-frame and the outcomes data-frame, as expected inputs for hardhat::mold() function.

Examples

library(dplyr)
library(data.tree)
data(starwars)
starwars_tree <- starwars %>%
  mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/")) %>%
  as.Node()
node_to_df(starwars_tree)
#> $x
#>    birth_year     eye_color
#> 1        19.0          blue
#> 2        41.9        yellow
#> 3        52.0          blue
#> 4        47.0          blue
#> 5        24.0         brown
#> 6        41.9          blue
#> 7        72.0         brown
#> 8        82.0          blue
#> 9        19.0         brown
#> 10       67.0         brown
#> 11         NA         brown
#> 12       57.0     blue-gray
#> 13       64.0          blue
#> 14       29.0         brown
#> 15       21.0         hazel
#> 16         NA          blue
#> 17       82.0        yellow
#> 18         NA         brown
#> 19         NA         brown
#> 20         NA         brown
#> 21       46.0         brown
#> 22       31.5         brown
#> 23       31.0         brown
#> 24       37.0          blue
#> 25       48.0          blue
#> 26         NA         brown
#> 27       92.0          blue
#> 28         NA          dark
#> 29         NA         hazel
#> 30         NA         brown
#> 31       91.0          blue
#> 32         NA          blue
#> 33       72.0         brown
#> 34      102.0         brown
#> 35       66.0         brown
#> 36      112.0        yellow
#> 37         NA           red
#> 38       33.0           red
#> 39       15.0           red
#> 40         NA     red, blue
#> 41         NA         black
#> 42      200.0          blue
#> 43         NA          blue
#> 44       44.0         black
#> 45      600.0        orange
#> 46      896.0         brown
#> 47       53.0           red
#> 48       41.0        orange
#> 49        8.0         brown
#> 50         NA         black
#> 51         NA           red
#> 52       52.0        orange
#> 53         NA        orange
#> 54         NA        orange
#> 55         NA          blue
#> 56       62.0         brown
#> 57         NA         white
#> 58         NA       unknown
#> 59         NA        yellow
#> 60         NA        orange
#> 61       54.0        yellow
#> 62         NA         brown
#> 63         NA          pink
#> 64       48.0         hazel
#> 65         NA        yellow
#> 66         NA         black
#> 67         NA        orange
#> 68       92.0        yellow
#> 69         NA         black
#> 70         NA          blue
#> 71         NA        orange
#> 72         NA        yellow
#> 73       22.0         black
#> 74         NA          blue
#> 75         NA        yellow
#> 76       58.0          blue
#> 77       40.0          blue
#> 78         NA        yellow
#> 79         NA        yellow
#> 80         NA         black
#> 81         NA         black
#> 82         NA       unknown
#> 83         NA       unknown
#> 84         NA          gold
#> 85         NA         black
#> 86         NA green, yellow
#> 87         NA         black
#>                                                                                                                                        films
#> 1                                            The Empire Strikes Back, Revenge of the Sith, Return of the Jedi, A New Hope, The Force Awakens
#> 2                                                               The Empire Strikes Back, Revenge of the Sith, Return of the Jedi, A New Hope
#> 3                                                                                      Attack of the Clones, Revenge of the Sith, A New Hope
#> 4                                                                                      Attack of the Clones, Revenge of the Sith, A New Hope
#> 5                                                                                                                                 A New Hope
#> 6                                                                              Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 7                                                                                                   Attack of the Clones, The Phantom Menace
#> 8                                                                                                                       Attack of the Clones
#> 9                                            The Empire Strikes Back, Revenge of the Sith, Return of the Jedi, A New Hope, The Force Awakens
#> 10                                                                                                 Attack of the Clones, Revenge of the Sith
#> 11                                                                                                           Revenge of the Sith, A New Hope
#> 12                    The Empire Strikes Back, Attack of the Clones, The Phantom Menace, Revenge of the Sith, Return of the Jedi, A New Hope
#> 13                                                                                                           Revenge of the Sith, A New Hope
#> 14                                                                The Empire Strikes Back, Return of the Jedi, A New Hope, The Force Awakens
#> 15                                                                                   The Empire Strikes Back, Return of the Jedi, A New Hope
#> 16                                                                                                                                A New Hope
#> 17                                The Empire Strikes Back, Attack of the Clones, The Phantom Menace, Revenge of the Sith, Return of the Jedi
#> 18                                                                                                                      Attack of the Clones
#> 19                                                                                                                      Attack of the Clones
#> 20                                                                                                                      Attack of the Clones
#> 21                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 22                                                                         The Empire Strikes Back, Attack of the Clones, Return of the Jedi
#> 23                                                                                               The Empire Strikes Back, Return of the Jedi
#> 24                                                                                                                   The Empire Strikes Back
#> 25                                                                                                                        Return of the Jedi
#> 26                                                                                                                        Return of the Jedi
#> 27                                                                                                                        The Phantom Menace
#> 28                                                                                                                         The Force Awakens
#> 29                                                                                                                         The Force Awakens
#> 30                                                                                                                         The Force Awakens
#> 31                                                                                                                        The Phantom Menace
#> 32                                                                                                                      Attack of the Clones
#> 33                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 34                                                                                                 Attack of the Clones, Revenge of the Sith
#> 35                                                                                                                      Attack of the Clones
#> 36                    The Empire Strikes Back, Attack of the Clones, The Phantom Menace, Revenge of the Sith, Return of the Jedi, A New Hope
#> 37                                                                                                                                A New Hope
#> 38 The Empire Strikes Back, Attack of the Clones, The Phantom Menace, Revenge of the Sith, Return of the Jedi, A New Hope, The Force Awakens
#> 39                                                                                                                   The Empire Strikes Back
#> 40                                                                                                 Attack of the Clones, Revenge of the Sith
#> 41                                                                                                                         The Force Awakens
#> 42                                           The Empire Strikes Back, Revenge of the Sith, Return of the Jedi, A New Hope, The Force Awakens
#> 43                                                                                                                       Revenge of the Sith
#> 44                                                                                                                                A New Hope
#> 45                                                                                        The Phantom Menace, Return of the Jedi, A New Hope
#> 46                                The Empire Strikes Back, Attack of the Clones, The Phantom Menace, Revenge of the Sith, Return of the Jedi
#> 47                                                                                                                   The Empire Strikes Back
#> 48                                                                                                     Return of the Jedi, The Force Awakens
#> 49                                                                                                                        Return of the Jedi
#> 50                                                                                                                        Return of the Jedi
#> 51                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 52                                                                                                  Attack of the Clones, The Phantom Menace
#> 53                                                                                                                        The Phantom Menace
#> 54                                                                                                                        The Phantom Menace
#> 55                                                                                                                        The Phantom Menace
#> 56                                                                                                                        The Phantom Menace
#> 57                                                                                                 Attack of the Clones, Revenge of the Sith
#> 58                                                                                                                         The Force Awakens
#> 59                                                                                                  Attack of the Clones, The Phantom Menace
#> 60                                                                                                                        The Phantom Menace
#> 61                                                                                                                        The Phantom Menace
#> 62                                                                                                   The Phantom Menace, Revenge of the Sith
#> 63                                                                                                                        Return of the Jedi
#> 64                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 65                                                                                                                        The Phantom Menace
#> 66                                                                                                                        The Phantom Menace
#> 67                                                                                                                        The Phantom Menace
#> 68                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 69                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 70                                                                                                   The Phantom Menace, Revenge of the Sith
#> 71                                                                                                   The Phantom Menace, Revenge of the Sith
#> 72                                                                                                                        The Phantom Menace
#> 73                                                                             Attack of the Clones, The Phantom Menace, Revenge of the Sith
#> 74                                                                                                  Attack of the Clones, The Phantom Menace
#> 75                                                                                                 Attack of the Clones, Revenge of the Sith
#> 76                                                                                                 Attack of the Clones, Revenge of the Sith
#> 77                                                                                                                      Attack of the Clones
#> 78                                                                                                                      Attack of the Clones
#> 79                                                                                                                      Attack of the Clones
#> 80                                                                                                                      Attack of the Clones
#> 81                                                                                                                      Attack of the Clones
#> 82                                                                                                                        The Phantom Menace
#> 83                                                                                                                      Attack of the Clones
#> 84                                                                                                                      Attack of the Clones
#> 85                                                                                                 Attack of the Clones, Revenge of the Sith
#> 86                                                                                                                       Revenge of the Sith
#> 87                                                                                                                       Revenge of the Sith
#>       gender    hair_color      homeworld   mass            sex
#> 1  masculine         blond       Tatooine   77.0           male
#> 2  masculine          none       Tatooine  136.0           male
#> 3  masculine   brown, grey       Tatooine  120.0           male
#> 4   feminine         brown       Tatooine   75.0         female
#> 5  masculine         black       Tatooine   84.0           male
#> 6  masculine         blond       Tatooine   84.0           male
#> 7   feminine         black       Tatooine     NA         female
#> 8  masculine         brown       Tatooine     NA           male
#> 9   feminine         brown       Alderaan   49.0         female
#> 10 masculine         black       Alderaan     NA           male
#> 11 masculine         brown       Alderaan   79.0           male
#> 12 masculine auburn, white        Stewjon   77.0           male
#> 13 masculine  auburn, grey         Eriadu     NA           male
#> 14 masculine         brown       Corellia   80.0           male
#> 15 masculine         brown       Corellia   77.0           male
#> 16 masculine         brown     Bestine IV  110.0           male
#> 17 masculine          grey          Naboo   75.0           male
#> 18 masculine         black          Naboo   85.0           male
#> 19  feminine         brown          Naboo     NA         female
#> 20  feminine         brown          Naboo     NA         female
#> 21  feminine         brown          Naboo   45.0         female
#> 22 masculine         black         Kamino   78.2           male
#> 23 masculine         black        Socorro   79.0           male
#> 24 masculine          none         Bespin   79.0           male
#> 25  feminine        auburn      Chandrila     NA         female
#> 26 masculine         brown           <NA>     NA           male
#> 27 masculine         brown           <NA>   89.0           male
#> 28 masculine         black           <NA>     NA           male
#> 29  feminine         brown           <NA>     NA         female
#> 30 masculine         brown           <NA>     NA           male
#> 31 masculine         blond      Coruscant     NA           male
#> 32  feminine         white      Coruscant     NA         female
#> 33 masculine          none     Haruun Kal   84.0           male
#> 34 masculine         white        Serenno   80.0           male
#> 35 masculine         black   Concord Dawn   79.0           male
#> 36 masculine          <NA>       Tatooine   75.0           none
#> 37 masculine          <NA>       Tatooine   32.0           none
#> 38 masculine          <NA>          Naboo   32.0           none
#> 39 masculine          none           <NA>  140.0           none
#> 40  feminine          none           <NA>     NA           none
#> 41 masculine          none           <NA>     NA           none
#> 42 masculine         brown       Kashyyyk  112.0           male
#> 43 masculine         brown       Kashyyyk  136.0           male
#> 44 masculine          <NA>          Rodia   74.0           male
#> 45 masculine          <NA>      Nal Hutta 1358.0 hermaphroditic
#> 46 masculine         white           <NA>   17.0           male
#> 47 masculine          none      Trandosha  113.0           male
#> 48 masculine          none       Mon Cala   83.0           male
#> 49 masculine         brown          Endor   20.0           male
#> 50 masculine          none        Sullust   68.0           male
#> 51 masculine          none Cato Neimoidia   90.0           male
#> 52 masculine          none          Naboo   66.0           male
#> 53 masculine          none          Naboo   82.0           male
#> 54 masculine          none          Naboo     NA           male
#> 55      <NA>         brown          Naboo     NA           <NA>
#> 56      <NA>         black          Naboo     NA           <NA>
#> 57      <NA>          none         Umbara   48.0           <NA>
#> 58      <NA>       unknown           <NA>     NA           <NA>
#> 59 masculine         black       Toydaria     NA           male
#> 60 masculine          none      Malastare   40.0           male
#> 61 masculine          none       Dathomir   80.0           male
#> 62 masculine         black       Iridonia     NA           male
#> 63 masculine          none         Ryloth     NA           male
#> 64  feminine          none         Ryloth   55.0         female
#> 65 masculine          none        Vulpter   45.0           male
#> 66 masculine          none        Troiken     NA           male
#> 67 masculine          none           Tund   65.0           male
#> 68 masculine         white          Cerea   82.0           male
#> 69 masculine          none    Glee Anselm   87.0           male
#> 70  feminine          none      Coruscant   50.0         female
#> 71 masculine          none        Iktotch     NA           male
#> 72 masculine          none        Quermia     NA           male
#> 73 masculine          none          Dorin   80.0           male
#> 74 masculine          none       Champala     NA           male
#> 75 masculine          none       Geonosis   80.0           male
#> 76  feminine         black         Mirial   56.2         female
#> 77  feminine         black         Mirial   50.0         female
#> 78  feminine        blonde          Zolan   55.0         female
#> 79 masculine          none           Ojom  102.0           male
#> 80 masculine          none         Kamino   88.0           male
#> 81  feminine          none         Kamino     NA         female
#> 82 masculine          none    Aleen Minor   15.0           male
#> 83 masculine          none          Skako   48.0           male
#> 84 masculine          none     Muunilinst     NA           male
#> 85  feminine          none          Shili   57.0         female
#> 86 masculine          none          Kalee  159.0           male
#> 87 masculine          none         Utapau   80.0           male
#>             skin_color        species
#> 1                 fair          Human
#> 2                white          Human
#> 3                light          Human
#> 4                light          Human
#> 5                light          Human
#> 6                 fair          Human
#> 7                 fair          Human
#> 8                 fair          Human
#> 9                light          Human
#> 10                 tan          Human
#> 11               light          Human
#> 12                fair          Human
#> 13                fair          Human
#> 14                fair          Human
#> 15                fair          Human
#> 16                fair          Human
#> 17                pale          Human
#> 18                dark          Human
#> 19               light          Human
#> 20               light          Human
#> 21               light          Human
#> 22                fair          Human
#> 23                dark          Human
#> 24               light          Human
#> 25                fair          Human
#> 26                fair          Human
#> 27                fair          Human
#> 28                dark          Human
#> 29               light          Human
#> 30               light          Human
#> 31                fair          Human
#> 32                fair          Human
#> 33                dark          Human
#> 34                fair          Human
#> 35                 tan          Human
#> 36                gold          Droid
#> 37          white, red          Droid
#> 38         white, blue          Droid
#> 39               metal          Droid
#> 40         silver, red          Droid
#> 41                none          Droid
#> 42             unknown        Wookiee
#> 43               brown        Wookiee
#> 44               green         Rodian
#> 45    green-tan, brown           Hutt
#> 46               green Yoda's species
#> 47               green     Trandoshan
#> 48        brown mottle   Mon Calamari
#> 49               brown           Ewok
#> 50                grey      Sullustan
#> 51       mottled green      Neimodian
#> 52              orange         Gungan
#> 53                grey         Gungan
#> 54               green         Gungan
#> 55                fair           <NA>
#> 56                dark           <NA>
#> 57                pale           <NA>
#> 58             unknown           <NA>
#> 59          blue, grey      Toydarian
#> 60           grey, red            Dug
#> 61                 red         Zabrak
#> 62               brown         Zabrak
#> 63                pale        Twi'lek
#> 64                blue        Twi'lek
#> 65          blue, grey     Vulptereen
#> 66         white, blue          Xexto
#> 67 grey, green, yellow          Toong
#> 68                pale         Cerean
#> 69               green       Nautolan
#> 70                dark     Tholothian
#> 71                pale       Iktotchi
#> 72               white       Quermian
#> 73              orange        Kel Dor
#> 74                blue       Chagrian
#> 75               green      Geonosian
#> 76              yellow       Mirialan
#> 77              yellow       Mirialan
#> 78 fair, green, yellow       Clawdite
#> 79               brown       Besalisk
#> 80                grey       Kaminoan
#> 81                grey       Kaminoan
#> 82          grey, blue         Aleena
#> 83         green, grey        Skakoan
#> 84                grey           Muun
#> 85    red, blue, white        Togruta
#> 86        brown, white        Kaleesh
#> 87                grey         Pau'an
#>                                                                                                   starships
#> 1                                                                                  X-wing, Imperial shuttle
#> 2                                                                                           TIE Advanced x1
#> 3                                                                                                      <NA>
#> 4                                                                                                      <NA>
#> 5                                                                                                    X-wing
#> 6                                                 Trade Federation cruiser, Jedi Interceptor, Naboo fighter
#> 7                                                                                                      <NA>
#> 8                                                                                                      <NA>
#> 9                                                                                                      <NA>
#> 10                                                                                                     <NA>
#> 11                                                                                                     <NA>
#> 12 Jedi starfighter, Trade Federation cruiser, Naboo star skiff, Jedi Interceptor, Belbullab-22 starfighter
#> 13                                                                                                     <NA>
#> 14                                                                      Millennium Falcon, Imperial shuttle
#> 15                                                                                                   X-wing
#> 16                                                                                                   X-wing
#> 17                                                                                                     <NA>
#> 18                                                                                            Naboo fighter
#> 19                                                                                                     <NA>
#> 20                                                                                                     <NA>
#> 21                                                     H-type Nubian yacht, Naboo star skiff, Naboo fighter
#> 22                                                                                                  Slave 1
#> 23                                                                                        Millennium Falcon
#> 24                                                                                                     <NA>
#> 25                                                                                                     <NA>
#> 26                                                                                                   A-wing
#> 27                                                                                                     <NA>
#> 28                                                                                                     <NA>
#> 29                                                                                                     <NA>
#> 30                                                                                      T-70 X-wing fighter
#> 31                                                                                                     <NA>
#> 32                                                                                                     <NA>
#> 33                                                                                                     <NA>
#> 34                                                                                                     <NA>
#> 35                                                                                                     <NA>
#> 36                                                                                                     <NA>
#> 37                                                                                                     <NA>
#> 38                                                                                                     <NA>
#> 39                                                                                                     <NA>
#> 40                                                                                                     <NA>
#> 41                                                                                                     <NA>
#> 42                                                                      Millennium Falcon, Imperial shuttle
#> 43                                                                                                     <NA>
#> 44                                                                                                     <NA>
#> 45                                                                                                     <NA>
#> 46                                                                                                     <NA>
#> 47                                                                                                     <NA>
#> 48                                                                                                     <NA>
#> 49                                                                                                     <NA>
#> 50                                                                                        Millennium Falcon
#> 51                                                                                                     <NA>
#> 52                                                                                                     <NA>
#> 53                                                                                                     <NA>
#> 54                                                                                                     <NA>
#> 55                                                                                     Naboo Royal Starship
#> 56                                                                                                     <NA>
#> 57                                                                                                     <NA>
#> 58                                                                                                     <NA>
#> 59                                                                                                     <NA>
#> 60                                                                                                     <NA>
#> 61                                                                                                 Scimitar
#> 62                                                                                                     <NA>
#> 63                                                                                                     <NA>
#> 64                                                                                                     <NA>
#> 65                                                                                                     <NA>
#> 66                                                                                                     <NA>
#> 67                                                                                                     <NA>
#> 68                                                                                                     <NA>
#> 69                                                                                                     <NA>
#> 70                                                                                                     <NA>
#> 71                                                                                                     <NA>
#> 72                                                                                                     <NA>
#> 73                                                                                         Jedi starfighter
#> 74                                                                                                     <NA>
#> 75                                                                                                     <NA>
#> 76                                                                                                     <NA>
#> 77                                                                                                     <NA>
#> 78                                                                                                     <NA>
#> 79                                                                                                     <NA>
#> 80                                                                                                     <NA>
#> 81                                                                                                     <NA>
#> 82                                                                                                     <NA>
#> 83                                                                                                     <NA>
#> 84                                                                                                     <NA>
#> 85                                                                                                     <NA>
#> 86                                                                                 Belbullab-22 starfighter
#> 87                                                                                                     <NA>
#>                                vehicles
#> 1    Snowspeeder, Imperial Speeder Bike
#> 2                                  <NA>
#> 3                                  <NA>
#> 4                                  <NA>
#> 5                                  <NA>
#> 6  Zephyr-G swoop bike, XJ-6 airspeeder
#> 7                                  <NA>
#> 8                                  <NA>
#> 9                 Imperial Speeder Bike
#> 10                                 <NA>
#> 11                                 <NA>
#> 12                      Tribubble bongo
#> 13                                 <NA>
#> 14                                 <NA>
#> 15                          Snowspeeder
#> 16                                 <NA>
#> 17                                 <NA>
#> 18                                 <NA>
#> 19                                 <NA>
#> 20                                 <NA>
#> 21                                 <NA>
#> 22                                 <NA>
#> 23                                 <NA>
#> 24                                 <NA>
#> 25                                 <NA>
#> 26                                 <NA>
#> 27                      Tribubble bongo
#> 28                                 <NA>
#> 29                                 <NA>
#> 30                                 <NA>
#> 31                                 <NA>
#> 32                                 <NA>
#> 33                                 <NA>
#> 34                     Flitknot speeder
#> 35                                 <NA>
#> 36                                 <NA>
#> 37                                 <NA>
#> 38                                 <NA>
#> 39                                 <NA>
#> 40                                 <NA>
#> 41                                 <NA>
#> 42                                AT-ST
#> 43                                 <NA>
#> 44                                 <NA>
#> 45                                 <NA>
#> 46                                 <NA>
#> 47                                 <NA>
#> 48                                 <NA>
#> 49                                 <NA>
#> 50                                 <NA>
#> 51                                 <NA>
#> 52                                 <NA>
#> 53                                 <NA>
#> 54                                 <NA>
#> 55                                 <NA>
#> 56                                 <NA>
#> 57                                 <NA>
#> 58                                 <NA>
#> 59                                 <NA>
#> 60                                 <NA>
#> 61                         Sith speeder
#> 62                                 <NA>
#> 63                                 <NA>
#> 64                                 <NA>
#> 65                                 <NA>
#> 66                                 <NA>
#> 67                                 <NA>
#> 68                                 <NA>
#> 69                                 <NA>
#> 70                                 <NA>
#> 71                                 <NA>
#> 72                                 <NA>
#> 73                                 <NA>
#> 74                                 <NA>
#> 75                                 <NA>
#> 76                                 <NA>
#> 77                                 <NA>
#> 78           Koro-2 Exodrive airspeeder
#> 79                                 <NA>
#> 80                                 <NA>
#> 81                                 <NA>
#> 82                                 <NA>
#> 83                                 <NA>
#> 84                                 <NA>
#> 85                                 <NA>
#> 86          Tsmeu-6 personal wheel bike
#> 87                                 <NA>
#> 
#> $y
#>           level_2        level_3
#> 1           Human       Tatooine
#> 2           Human       Tatooine
#> 3           Human       Tatooine
#> 4           Human       Tatooine
#> 5           Human       Tatooine
#> 6           Human       Tatooine
#> 7           Human       Tatooine
#> 8           Human       Tatooine
#> 9           Human       Alderaan
#> 10          Human       Alderaan
#> 11          Human       Alderaan
#> 12          Human        Stewjon
#> 13          Human         Eriadu
#> 14          Human       Corellia
#> 15          Human       Corellia
#> 16          Human     Bestine IV
#> 17          Human          Naboo
#> 18          Human          Naboo
#> 19          Human          Naboo
#> 20          Human          Naboo
#> 21          Human          Naboo
#> 22          Human         Kamino
#> 23          Human        Socorro
#> 24          Human         Bespin
#> 25          Human      Chandrila
#> 26          Human             NA
#> 27          Human             NA
#> 28          Human             NA
#> 29          Human             NA
#> 30          Human             NA
#> 31          Human      Coruscant
#> 32          Human      Coruscant
#> 33          Human     Haruun Kal
#> 34          Human        Serenno
#> 35          Human   Concord Dawn
#> 36          Droid       Tatooine
#> 37          Droid       Tatooine
#> 38          Droid          Naboo
#> 39          Droid             NA
#> 40          Droid             NA
#> 41          Droid             NA
#> 42        Wookiee       Kashyyyk
#> 43        Wookiee       Kashyyyk
#> 44         Rodian          Rodia
#> 45           Hutt      Nal Hutta
#> 46 Yoda's species             NA
#> 47     Trandoshan      Trandosha
#> 48   Mon Calamari       Mon Cala
#> 49           Ewok          Endor
#> 50      Sullustan        Sullust
#> 51      Neimodian Cato Neimoidia
#> 52         Gungan          Naboo
#> 53         Gungan          Naboo
#> 54         Gungan          Naboo
#> 55             NA          Naboo
#> 56             NA          Naboo
#> 57             NA         Umbara
#> 58             NA             NA
#> 59      Toydarian       Toydaria
#> 60            Dug      Malastare
#> 61         Zabrak       Dathomir
#> 62         Zabrak       Iridonia
#> 63        Twi'lek         Ryloth
#> 64        Twi'lek         Ryloth
#> 65     Vulptereen        Vulpter
#> 66          Xexto        Troiken
#> 67          Toong           Tund
#> 68         Cerean          Cerea
#> 69       Nautolan    Glee Anselm
#> 70     Tholothian      Coruscant
#> 71       Iktotchi        Iktotch
#> 72       Quermian        Quermia
#> 73        Kel Dor          Dorin
#> 74       Chagrian       Champala
#> 75      Geonosian       Geonosis
#> 76       Mirialan         Mirial
#> 77       Mirialan         Mirial
#> 78       Clawdite          Zolan
#> 79       Besalisk           Ojom
#> 80       Kaminoan         Kamino
#> 81       Kaminoan         Kamino
#> 82         Aleena    Aleen Minor
#> 83        Skakoan          Skako
#> 84           Muun     Muunilinst
#> 85        Togruta          Shili
#> 86        Kaleesh          Kalee
#> 87         Pau'an         Utapau
#>