Turn a Node object into predictor and outcome.

Turn a Node object into predictor and outcome.

node_to_df(x, drop_last_level = TRUE)

Arguments

x

Node object

drop_last_level

TRUE unused

Value

a named list of x and y, being respectively the predictor data-frame and the outcomes data-frame, as expected inputs for hardhat::mold() function.

Examples

library(dplyr)
library(data.tree)
data(starwars)
starwars_tree <- starwars %>%
  mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/")) %>%
  as.Node()
node_to_df(starwars_tree)
#> $x
#>    birth_year     eye_color
#> 1        19.0          blue
#> 2        41.9        yellow
#> 3        52.0          blue
#> 4        47.0          blue
#> 5        24.0         brown
#> 6        41.9          blue
#> 7        72.0         brown
#> 8        82.0          blue
#> 9        19.0         brown
#> 10       67.0         brown
#> 11         NA         brown
#> 12       57.0     blue-gray
#> 13       64.0          blue
#> 14       29.0         brown
#> 15       21.0         hazel
#> 16       82.0        yellow
#> 17       46.0         brown
#> 18         NA          blue
#> 19       62.0         brown
#> 20         NA         brown
#> 21       31.5         brown
#> 22       31.0         brown
#> 23       37.0          blue
#> 24       48.0          blue
#> 25         NA         brown
#> 26       92.0          blue
#> 27         NA          dark
#> 28         NA         hazel
#> 29         NA         brown
#> 30         NA       unknown
#> 31       91.0          blue
#> 32         NA          blue
#> 33       72.0         brown
#> 34      102.0         brown
#> 35       66.0         brown
#> 36      112.0        yellow
#> 37         NA           red
#> 38       33.0           red
#> 39       15.0           red
#> 40         NA     red, blue
#> 41         NA         black
#> 42      200.0          blue
#> 43         NA          blue
#> 44       44.0         black
#> 45      600.0        orange
#> 46         NA          blue
#> 47         NA         brown
#> 48         NA         brown
#> 49         NA         white
#> 50      896.0         brown
#> 51       53.0           red
#> 52       41.0        orange
#> 53        8.0         brown
#> 54         NA         black
#> 55         NA           red
#> 56       52.0        orange
#> 57         NA        orange
#> 58         NA        orange
#> 59         NA        yellow
#> 60         NA        orange
#> 61       54.0        yellow
#> 62         NA         brown
#> 63         NA          pink
#> 64       48.0         hazel
#> 65         NA       unknown
#> 66         NA        yellow
#> 67         NA         black
#> 68         NA        orange
#> 69       92.0        yellow
#> 70         NA         black
#> 71         NA          blue
#> 72         NA        orange
#> 73         NA        yellow
#> 74       22.0         black
#> 75         NA          blue
#> 76         NA        yellow
#> 77       58.0          blue
#> 78       40.0          blue
#> 79         NA        yellow
#> 80         NA        yellow
#> 81         NA         black
#> 82         NA         black
#> 83         NA       unknown
#> 84         NA          gold
#> 85         NA         black
#> 86         NA green, yellow
#> 87         NA         black
#>                                                                                                                                        films
#> 1                                            A New Hope, The Empire Strikes Back, Return of the Jedi, Revenge of the Sith, The Force Awakens
#> 2                                                               A New Hope, The Empire Strikes Back, Return of the Jedi, Revenge of the Sith
#> 3                                                                                      A New Hope, Attack of the Clones, Revenge of the Sith
#> 4                                                                                      A New Hope, Attack of the Clones, Revenge of the Sith
#> 5                                                                                                                                 A New Hope
#> 6                                                                              The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 7                                                                                                   The Phantom Menace, Attack of the Clones
#> 8                                                                                                                       Attack of the Clones
#> 9                                            A New Hope, The Empire Strikes Back, Return of the Jedi, Revenge of the Sith, The Force Awakens
#> 10                                                                                                 Attack of the Clones, Revenge of the Sith
#> 11                                                                                                           A New Hope, Revenge of the Sith
#> 12                    A New Hope, The Empire Strikes Back, Return of the Jedi, The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 13                                                                                                           A New Hope, Revenge of the Sith
#> 14                                                                A New Hope, The Empire Strikes Back, Return of the Jedi, The Force Awakens
#> 15                                                                                   A New Hope, The Empire Strikes Back, Return of the Jedi
#> 16                                The Empire Strikes Back, Return of the Jedi, The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 17                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 18                                                                                                                        The Phantom Menace
#> 19                                                                                                                        The Phantom Menace
#> 20                                                                                                                      Attack of the Clones
#> 21                                                                         The Empire Strikes Back, Return of the Jedi, Attack of the Clones
#> 22                                                                                               The Empire Strikes Back, Return of the Jedi
#> 23                                                                                                                   The Empire Strikes Back
#> 24                                                                                                                        Return of the Jedi
#> 25                                                                                                                        Return of the Jedi
#> 26                                                                                                                        The Phantom Menace
#> 27                                                                                                                         The Force Awakens
#> 28                                                                                                                         The Force Awakens
#> 29                                                                                                                         The Force Awakens
#> 30                                                                                                                         The Force Awakens
#> 31                                                                                                                        The Phantom Menace
#> 32                                                                                                                      Attack of the Clones
#> 33                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 34                                                                                                 Attack of the Clones, Revenge of the Sith
#> 35                                                                                                                      Attack of the Clones
#> 36                    A New Hope, The Empire Strikes Back, Return of the Jedi, The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 37                                                                                                                                A New Hope
#> 38 A New Hope, The Empire Strikes Back, Return of the Jedi, The Phantom Menace, Attack of the Clones, Revenge of the Sith, The Force Awakens
#> 39                                                                                                                   The Empire Strikes Back
#> 40                                                                                                 Attack of the Clones, Revenge of the Sith
#> 41                                                                                                                         The Force Awakens
#> 42                                           A New Hope, The Empire Strikes Back, Return of the Jedi, Revenge of the Sith, The Force Awakens
#> 43                                                                                                                       Revenge of the Sith
#> 44                                                                                                                                A New Hope
#> 45                                                                                        A New Hope, Return of the Jedi, The Phantom Menace
#> 46                                                                                                                                A New Hope
#> 47                                                                                                                      Attack of the Clones
#> 48                                                                                                                      Attack of the Clones
#> 49                                                                                                 Attack of the Clones, Revenge of the Sith
#> 50                                The Empire Strikes Back, Return of the Jedi, The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 51                                                                                                                   The Empire Strikes Back
#> 52                                                                                                     Return of the Jedi, The Force Awakens
#> 53                                                                                                                        Return of the Jedi
#> 54                                                                                                                        Return of the Jedi
#> 55                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 56                                                                                                  The Phantom Menace, Attack of the Clones
#> 57                                                                                                                        The Phantom Menace
#> 58                                                                                                                        The Phantom Menace
#> 59                                                                                                  The Phantom Menace, Attack of the Clones
#> 60                                                                                                                        The Phantom Menace
#> 61                                                                                                                        The Phantom Menace
#> 62                                                                                                   The Phantom Menace, Revenge of the Sith
#> 63                                                                                                                        Return of the Jedi
#> 64                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 65                                                                                                                        The Phantom Menace
#> 66                                                                                                                        The Phantom Menace
#> 67                                                                                                                        The Phantom Menace
#> 68                                                                                                                        The Phantom Menace
#> 69                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 70                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 71                                                                                                   The Phantom Menace, Revenge of the Sith
#> 72                                                                                                   The Phantom Menace, Revenge of the Sith
#> 73                                                                                                                        The Phantom Menace
#> 74                                                                             The Phantom Menace, Attack of the Clones, Revenge of the Sith
#> 75                                                                                                  The Phantom Menace, Attack of the Clones
#> 76                                                                                                 Attack of the Clones, Revenge of the Sith
#> 77                                                                                                 Attack of the Clones, Revenge of the Sith
#> 78                                                                                                                      Attack of the Clones
#> 79                                                                                                                      Attack of the Clones
#> 80                                                                                                                      Attack of the Clones
#> 81                                                                                                                      Attack of the Clones
#> 82                                                                                                                      Attack of the Clones
#> 83                                                                                                                      Attack of the Clones
#> 84                                                                                                                      Attack of the Clones
#> 85                                                                                                 Attack of the Clones, Revenge of the Sith
#> 86                                                                                                                       Revenge of the Sith
#> 87                                                                                                                       Revenge of the Sith
#>       gender    hair_color      homeworld   mass            sex
#> 1  masculine         blond       Tatooine   77.0           male
#> 2  masculine          none       Tatooine  136.0           male
#> 3  masculine   brown, grey       Tatooine  120.0           male
#> 4   feminine         brown       Tatooine   75.0         female
#> 5  masculine         black       Tatooine   84.0           male
#> 6  masculine         blond       Tatooine   84.0           male
#> 7   feminine         black       Tatooine     NA         female
#> 8  masculine         brown       Tatooine     NA           male
#> 9   feminine         brown       Alderaan   49.0         female
#> 10 masculine         black       Alderaan     NA           male
#> 11 masculine         brown       Alderaan   79.0           male
#> 12 masculine auburn, white        Stewjon   77.0           male
#> 13 masculine  auburn, grey         Eriadu     NA           male
#> 14 masculine         brown       Corellia   80.0           male
#> 15 masculine         brown       Corellia   77.0           male
#> 16 masculine          grey          Naboo   75.0           male
#> 17  feminine         brown          Naboo   45.0         female
#> 18 masculine         brown          Naboo     NA           male
#> 19 masculine         black          Naboo     NA           male
#> 20  feminine         brown          Naboo     NA         female
#> 21 masculine         black         Kamino   78.2           male
#> 22 masculine         black        Socorro   79.0           male
#> 23 masculine          none         Bespin   79.0           male
#> 24  feminine        auburn      Chandrila     NA         female
#> 25 masculine         brown           <NA>     NA           male
#> 26 masculine         brown           <NA>   89.0           male
#> 27 masculine         black           <NA>     NA           male
#> 28  feminine         brown           <NA>     NA         female
#> 29 masculine         brown           <NA>     NA           male
#> 30  feminine          none           <NA>     NA         female
#> 31 masculine         blond      Coruscant     NA           male
#> 32  feminine         white      Coruscant     NA         female
#> 33 masculine          none     Haruun Kal   84.0           male
#> 34 masculine         white        Serenno   80.0           male
#> 35 masculine         black   Concord Dawn   79.0           male
#> 36 masculine          <NA>       Tatooine   75.0           none
#> 37 masculine          <NA>       Tatooine   32.0           none
#> 38 masculine          <NA>          Naboo   32.0           none
#> 39 masculine          none           <NA>  140.0           none
#> 40  feminine          none           <NA>     NA           none
#> 41 masculine          none           <NA>     NA           none
#> 42 masculine         brown       Kashyyyk  112.0           male
#> 43 masculine         brown       Kashyyyk  136.0           male
#> 44 masculine          <NA>          Rodia   74.0           male
#> 45 masculine          <NA>      Nal Hutta 1358.0 hermaphroditic
#> 46      <NA>         brown     Bestine IV  110.0           <NA>
#> 47      <NA>         black          Naboo   85.0           <NA>
#> 48      <NA>         brown          Naboo     NA           <NA>
#> 49      <NA>          none         Umbara   48.0           <NA>
#> 50 masculine         white           <NA>   17.0           male
#> 51 masculine          none      Trandosha  113.0           male
#> 52 masculine          none       Mon Cala   83.0           male
#> 53 masculine         brown          Endor   20.0           male
#> 54 masculine          none        Sullust   68.0           male
#> 55 masculine          none Cato Neimoidia   90.0           male
#> 56 masculine          none          Naboo   66.0           male
#> 57 masculine          none          Naboo   82.0           male
#> 58 masculine          none          Naboo     NA           male
#> 59 masculine         black       Toydaria     NA           male
#> 60 masculine          none      Malastare   40.0           male
#> 61 masculine          none       Dathomir   80.0           male
#> 62 masculine         black       Iridonia     NA           male
#> 63 masculine          none         Ryloth     NA           male
#> 64  feminine          none         Ryloth   55.0         female
#> 65 masculine          none    Aleen Minor   15.0           male
#> 66 masculine          none        Vulpter   45.0           male
#> 67 masculine          none        Troiken     NA           male
#> 68 masculine          none           Tund   65.0           male
#> 69 masculine         white          Cerea   82.0           male
#> 70 masculine          none    Glee Anselm   87.0           male
#> 71  feminine          none      Coruscant   50.0         female
#> 72 masculine          none        Iktotch     NA           male
#> 73 masculine          none        Quermia     NA           male
#> 74 masculine          none          Dorin   80.0           male
#> 75 masculine          none       Champala     NA           male
#> 76 masculine          none       Geonosis   80.0           male
#> 77  feminine         black         Mirial   56.2         female
#> 78  feminine         black         Mirial   50.0         female
#> 79  feminine        blonde          Zolan   55.0         female
#> 80 masculine          none           Ojom  102.0           male
#> 81 masculine          none         Kamino   88.0           male
#> 82  feminine          none         Kamino     NA         female
#> 83 masculine          none          Skako   48.0           male
#> 84 masculine          none     Muunilinst     NA           male
#> 85  feminine          none          Shili   57.0         female
#> 86 masculine          none          Kalee  159.0           male
#> 87 masculine          none         Utapau   80.0           male
#>             skin_color        species
#> 1                 fair          Human
#> 2                white          Human
#> 3                light          Human
#> 4                light          Human
#> 5                light          Human
#> 6                 fair          Human
#> 7                 fair          Human
#> 8                 fair          Human
#> 9                light          Human
#> 10                 tan          Human
#> 11               light          Human
#> 12                fair          Human
#> 13                fair          Human
#> 14                fair          Human
#> 15                fair          Human
#> 16                pale          Human
#> 17               light          Human
#> 18                fair          Human
#> 19                dark          Human
#> 20               light          Human
#> 21                fair          Human
#> 22                dark          Human
#> 23               light          Human
#> 24                fair          Human
#> 25                fair          Human
#> 26                fair          Human
#> 27                dark          Human
#> 28               light          Human
#> 29               light          Human
#> 30                none          Human
#> 31                fair          Human
#> 32                fair          Human
#> 33                dark          Human
#> 34                fair          Human
#> 35                 tan          Human
#> 36                gold          Droid
#> 37          white, red          Droid
#> 38         white, blue          Droid
#> 39               metal          Droid
#> 40         silver, red          Droid
#> 41                none          Droid
#> 42             unknown        Wookiee
#> 43               brown        Wookiee
#> 44               green         Rodian
#> 45    green-tan, brown           Hutt
#> 46                fair           <NA>
#> 47                dark           <NA>
#> 48               light           <NA>
#> 49                pale           <NA>
#> 50               green Yoda's species
#> 51               green     Trandoshan
#> 52        brown mottle   Mon Calamari
#> 53               brown           Ewok
#> 54                grey      Sullustan
#> 55       mottled green      Neimodian
#> 56              orange         Gungan
#> 57                grey         Gungan
#> 58               green         Gungan
#> 59          blue, grey      Toydarian
#> 60           grey, red            Dug
#> 61                 red         Zabrak
#> 62               brown         Zabrak
#> 63                pale        Twi'lek
#> 64                blue        Twi'lek
#> 65          grey, blue         Aleena
#> 66          blue, grey     Vulptereen
#> 67         white, blue          Xexto
#> 68 grey, green, yellow          Toong
#> 69                pale         Cerean
#> 70               green       Nautolan
#> 71                dark     Tholothian
#> 72                pale       Iktotchi
#> 73               white       Quermian
#> 74              orange        Kel Dor
#> 75                blue       Chagrian
#> 76               green      Geonosian
#> 77              yellow       Mirialan
#> 78              yellow       Mirialan
#> 79 fair, green, yellow       Clawdite
#> 80               brown       Besalisk
#> 81                grey       Kaminoan
#> 82                grey       Kaminoan
#> 83         green, grey        Skakoan
#> 84                grey           Muun
#> 85    red, blue, white        Togruta
#> 86        brown, white        Kaleesh
#> 87                grey         Pau'an
#>                                                                                                   starships
#> 1                                                                                  X-wing, Imperial shuttle
#> 2                                                                                           TIE Advanced x1
#> 3                                                                                                      <NA>
#> 4                                                                                                      <NA>
#> 5                                                                                                    X-wing
#> 6                                                 Naboo fighter, Trade Federation cruiser, Jedi Interceptor
#> 7                                                                                                      <NA>
#> 8                                                                                                      <NA>
#> 9                                                                                                      <NA>
#> 10                                                                                                     <NA>
#> 11                                                                                                     <NA>
#> 12 Jedi starfighter, Trade Federation cruiser, Naboo star skiff, Jedi Interceptor, Belbullab-22 starfighter
#> 13                                                                                                     <NA>
#> 14                                                                      Millennium Falcon, Imperial shuttle
#> 15                                                                                                   X-wing
#> 16                                                                                                     <NA>
#> 17                                                     Naboo fighter, H-type Nubian yacht, Naboo star skiff
#> 18                                                                                     Naboo Royal Starship
#> 19                                                                                                     <NA>
#> 20                                                                                                     <NA>
#> 21                                                                                                  Slave 1
#> 22                                                                                        Millennium Falcon
#> 23                                                                                                     <NA>
#> 24                                                                                                     <NA>
#> 25                                                                                                   A-wing
#> 26                                                                                                     <NA>
#> 27                                                                                                     <NA>
#> 28                                                                                                     <NA>
#> 29                                                                                                   X-wing
#> 30                                                                                                     <NA>
#> 31                                                                                                     <NA>
#> 32                                                                                                     <NA>
#> 33                                                                                                     <NA>
#> 34                                                                                                     <NA>
#> 35                                                                                                     <NA>
#> 36                                                                                                     <NA>
#> 37                                                                                                     <NA>
#> 38                                                                                                     <NA>
#> 39                                                                                                     <NA>
#> 40                                                                                                     <NA>
#> 41                                                                                                     <NA>
#> 42                                                                      Millennium Falcon, Imperial shuttle
#> 43                                                                                                     <NA>
#> 44                                                                                                     <NA>
#> 45                                                                                                     <NA>
#> 46                                                                                                   X-wing
#> 47                                                                                            Naboo fighter
#> 48                                                                                                     <NA>
#> 49                                                                                                     <NA>
#> 50                                                                                                     <NA>
#> 51                                                                                                     <NA>
#> 52                                                                                                     <NA>
#> 53                                                                                                     <NA>
#> 54                                                                                        Millennium Falcon
#> 55                                                                                                     <NA>
#> 56                                                                                                     <NA>
#> 57                                                                                                     <NA>
#> 58                                                                                                     <NA>
#> 59                                                                                                     <NA>
#> 60                                                                                                     <NA>
#> 61                                                                                                 Scimitar
#> 62                                                                                                     <NA>
#> 63                                                                                                     <NA>
#> 64                                                                                                     <NA>
#> 65                                                                                                     <NA>
#> 66                                                                                                     <NA>
#> 67                                                                                                     <NA>
#> 68                                                                                                     <NA>
#> 69                                                                                                     <NA>
#> 70                                                                                                     <NA>
#> 71                                                                                                     <NA>
#> 72                                                                                                     <NA>
#> 73                                                                                                     <NA>
#> 74                                                                                         Jedi starfighter
#> 75                                                                                                     <NA>
#> 76                                                                                                     <NA>
#> 77                                                                                                     <NA>
#> 78                                                                                                     <NA>
#> 79                                                                                                     <NA>
#> 80                                                                                                     <NA>
#> 81                                                                                                     <NA>
#> 82                                                                                                     <NA>
#> 83                                                                                                     <NA>
#> 84                                                                                                     <NA>
#> 85                                                                                                     <NA>
#> 86                                                                                 Belbullab-22 starfighter
#> 87                                                                                                     <NA>
#>                                vehicles
#> 1    Snowspeeder, Imperial Speeder Bike
#> 2                                  <NA>
#> 3                                  <NA>
#> 4                                  <NA>
#> 5                                  <NA>
#> 6  Zephyr-G swoop bike, XJ-6 airspeeder
#> 7                                  <NA>
#> 8                                  <NA>
#> 9                 Imperial Speeder Bike
#> 10                                 <NA>
#> 11                                 <NA>
#> 12                      Tribubble bongo
#> 13                                 <NA>
#> 14                                 <NA>
#> 15                          Snowspeeder
#> 16                                 <NA>
#> 17                                 <NA>
#> 18                                 <NA>
#> 19                                 <NA>
#> 20                                 <NA>
#> 21                                 <NA>
#> 22                                 <NA>
#> 23                                 <NA>
#> 24                                 <NA>
#> 25                                 <NA>
#> 26                      Tribubble bongo
#> 27                                 <NA>
#> 28                                 <NA>
#> 29                                 <NA>
#> 30                                 <NA>
#> 31                                 <NA>
#> 32                                 <NA>
#> 33                                 <NA>
#> 34                     Flitknot speeder
#> 35                                 <NA>
#> 36                                 <NA>
#> 37                                 <NA>
#> 38                                 <NA>
#> 39                                 <NA>
#> 40                                 <NA>
#> 41                                 <NA>
#> 42                                AT-ST
#> 43                                 <NA>
#> 44                                 <NA>
#> 45                                 <NA>
#> 46                                 <NA>
#> 47                                 <NA>
#> 48                                 <NA>
#> 49                                 <NA>
#> 50                                 <NA>
#> 51                                 <NA>
#> 52                                 <NA>
#> 53                                 <NA>
#> 54                                 <NA>
#> 55                                 <NA>
#> 56                                 <NA>
#> 57                                 <NA>
#> 58                                 <NA>
#> 59                                 <NA>
#> 60                                 <NA>
#> 61                         Sith speeder
#> 62                                 <NA>
#> 63                                 <NA>
#> 64                                 <NA>
#> 65                                 <NA>
#> 66                                 <NA>
#> 67                                 <NA>
#> 68                                 <NA>
#> 69                                 <NA>
#> 70                                 <NA>
#> 71                                 <NA>
#> 72                                 <NA>
#> 73                                 <NA>
#> 74                                 <NA>
#> 75                                 <NA>
#> 76                                 <NA>
#> 77                                 <NA>
#> 78                                 <NA>
#> 79           Koro-2 Exodrive airspeeder
#> 80                                 <NA>
#> 81                                 <NA>
#> 82                                 <NA>
#> 83                                 <NA>
#> 84                                 <NA>
#> 85                                 <NA>
#> 86          Tsmeu-6 personal wheel bike
#> 87                                 <NA>
#> 
#> $y
#>           level_2        level_3
#> 1           Human       Tatooine
#> 2           Human       Tatooine
#> 3           Human       Tatooine
#> 4           Human       Tatooine
#> 5           Human       Tatooine
#> 6           Human       Tatooine
#> 7           Human       Tatooine
#> 8           Human       Tatooine
#> 9           Human       Alderaan
#> 10          Human       Alderaan
#> 11          Human       Alderaan
#> 12          Human        Stewjon
#> 13          Human         Eriadu
#> 14          Human       Corellia
#> 15          Human       Corellia
#> 16          Human          Naboo
#> 17          Human          Naboo
#> 18          Human          Naboo
#> 19          Human          Naboo
#> 20          Human          Naboo
#> 21          Human         Kamino
#> 22          Human        Socorro
#> 23          Human         Bespin
#> 24          Human      Chandrila
#> 25          Human             NA
#> 26          Human             NA
#> 27          Human             NA
#> 28          Human             NA
#> 29          Human             NA
#> 30          Human             NA
#> 31          Human      Coruscant
#> 32          Human      Coruscant
#> 33          Human     Haruun Kal
#> 34          Human        Serenno
#> 35          Human   Concord Dawn
#> 36          Droid       Tatooine
#> 37          Droid       Tatooine
#> 38          Droid          Naboo
#> 39          Droid             NA
#> 40          Droid             NA
#> 41          Droid             NA
#> 42        Wookiee       Kashyyyk
#> 43        Wookiee       Kashyyyk
#> 44         Rodian          Rodia
#> 45           Hutt      Nal Hutta
#> 46             NA     Bestine IV
#> 47             NA          Naboo
#> 48             NA          Naboo
#> 49             NA         Umbara
#> 50 Yoda's species             NA
#> 51     Trandoshan      Trandosha
#> 52   Mon Calamari       Mon Cala
#> 53           Ewok          Endor
#> 54      Sullustan        Sullust
#> 55      Neimodian Cato Neimoidia
#> 56         Gungan          Naboo
#> 57         Gungan          Naboo
#> 58         Gungan          Naboo
#> 59      Toydarian       Toydaria
#> 60            Dug      Malastare
#> 61         Zabrak       Dathomir
#> 62         Zabrak       Iridonia
#> 63        Twi'lek         Ryloth
#> 64        Twi'lek         Ryloth
#> 65         Aleena    Aleen Minor
#> 66     Vulptereen        Vulpter
#> 67          Xexto        Troiken
#> 68          Toong           Tund
#> 69         Cerean          Cerea
#> 70       Nautolan    Glee Anselm
#> 71     Tholothian      Coruscant
#> 72       Iktotchi        Iktotch
#> 73       Quermian        Quermia
#> 74        Kel Dor          Dorin
#> 75       Chagrian       Champala
#> 76      Geonosian       Geonosis
#> 77       Mirialan         Mirial
#> 78       Mirialan         Mirial
#> 79       Clawdite          Zolan
#> 80       Besalisk           Ojom
#> 81       Kaminoan         Kamino
#> 82       Kaminoan         Kamino
#> 83        Skakoan          Skako
#> 84           Muun     Muunilinst
#> 85        Togruta          Shili
#> 86        Kaleesh          Kalee
#> 87         Pau'an         Utapau
#>